# openapi/glenkeos-complete-api-v1.yaml
openapi: 3.0.3
info:
  title: GlenKeos Platform API
  version: 1.0.0
  description: |
    Complete API specification for the GlenKeos multi-tenant commerce platform.
    Includes all 10 microservices, PayPal integration, and event-driven workflows.
  contact:
    name: GlenKeos Engineering
    email: engineering@glenkeos.com
servers:
  - url: https://api.glenkeos.com/v1
    description: Production
  - url: https://staging-api.glenkeos.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
security:
  - CognitoAuthorizer: []
components:
  securitySchemes:
    CognitoAuthorizer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        phone:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        reward_points:
          type: integer
        tier:
          type: string
          enum: [BRONZE, SILVER, GOLD, PLATINUM]
        created_at:
          type: string
          format: date-time

    # Orders
...
paths:
...
  /customers:
    post:
      summary: Create customer
      operationId: createCustomer
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, phone, first_name, last_name]
              properties:
                email:
                  type: string
                  format: email
                phone:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

    get:
      summary: List customers
      operationId: listCustomers
      tags: [Customers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
  /customers/{customerId}:
    get:
      summary: Get customer by ID
      operationId: getCustomer
      tags: [Customers]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
...